feat: Added the new functions in post ata

+++ sid/src/main/java/br/sc/weg/sid/model/entities/PropostaResumida.java
@@ -11,7 +11,7 @@ public class PropostaResumida {
 
     Integer tempoDeExecucaoDemanda;
 
-    Integer valorDemanda;
+    Double valorDemanda;
 
     Integer idDemanda;
